+++ library/alloc/src/string.rs
@@ -53,7 +53,7 @@ use core::ops::Add;
 #[cfg(not(no_global_oom_handling))]
 use core::ops::AddAssign;
 use core::ops::{self, Range, RangeBounds};
-use core::str::pattern::{Pattern, Utf8Pattern};
+use core::pattern::{Pattern, Utf8Pattern};
 use core::{fmt, hash, ptr, slice};
 
 #[cfg(not(no_global_oom_handling))]
@@ -1584,7 +1584,7 @@ impl String {
     where
         P: for<'x> Pattern<&'x str>,
     {
-        use core::str::pattern::Searcher;
+        use core::pattern::Searcher;
 
         let rejections = {
             let mut searcher = pat.into_searcher(self);
@@ -2164,7 +2164,7 @@ impl String {
     #[unstable(feature = "string_replace_in_place", issue = "147949")]
     pub fn replace_last<'a, P>(&'a mut self, from: P, to: &str)
     where
-        P: for<'x> Pattern<&'x str, Searcher: core::str::pattern::ReverseSearcher<&'x str>>,
+        P: for<'x> Pattern<&'x str, Searcher: core::pattern::ReverseSearcher<&'x str>>,
     {
         let range = match self.rmatch_indices(from).next() {
             Some((start, match_str)) => start..start + match_str.len(),
@@ -2682,7 +2682,7 @@ impl<'a, 'b> Pattern<&'a str> for &'b String {
     #[inline]
     fn is_suffix_of(self, haystack: &'a str) -> bool
     where
-        Self::Searcher: core::str::pattern::ReverseSearcher<&'a str>,
+        Self::Searcher: core::pattern::ReverseSearcher<&'a str>,
     {
         self[..].is_suffix_of(haystack)
     }
@@ -2690,7 +2690,7 @@ impl<'a, 'b> Pattern<&'a str> for &'b String {
     #[inline]
     fn strip_suffix_of(self, haystack: &'a str) -> Option<&'a str>
     where
-        Self::Searcher: core::str::pattern::ReverseSearcher<&'a str>,
+        Self::Searcher: core::pattern::ReverseSearcher<&'a str>,
     {
         self[..].strip_suffix_of(haystack)
     }
