+++ library/core/src/lib.rs
@@ -318,6 +318,7 @@ pub mod unsafe_binder;
 
 pub mod fmt;
 pub mod hash;
+pub mod pattern;
 pub mod slice;
 pub mod str;
 pub mod time;
